+++ src/cache/cache.rs
@@ -11,9 +11,6 @@ use serde::ser::Serialize;
 use serde_json;
 use std::time::Duration;
 
-pub const CACHE_RESP_PREFIX: &'static str = "c_resp";
-pub const CACHE_REQS_PREFIX: &'static str = "c_reqs";
-
 #[database("service_cache")]
 pub struct ServiceCache(redis::Connection);
 
@@ -71,95 +68,6 @@ impl Cache for ServiceCache {
         info(self)
     }
 }
-
-pub(super) trait CacheExt: Cache {
-    fn request_cached_op(
-        &self,
-        client: &reqwest::blocking::Client,
-        operation: &RequestCached,
-    ) -> ApiResult<String> {
-        self.request_cached_advanced(
-            client,
-            &operation.url,
-            operation.cache_duration,
-            operation.error_cache_duration,
-            operation.cache_all_errors,
-            operation.request_timeout,
-        )
-    }
-
-    fn request_cached_advanced(
-        &self,
-        client: &reqwest::blocking::Client,
-        url: &str,
-        cache_duration: usize,
-        error_cache_duration: usize,
-        cache_all_errors: bool,
-        request_timeout: u64,
-    ) -> ApiResult<String> {
-        let cache_key = format!("{}_{}", CACHE_REQS_PREFIX, &url);
-        match self.fetch(&cache_key) {
-            Some(cached) => CachedWithCode::split(&cached).to_result(),
-            None => {
-                let mut request = client.get(url);
-                if request_timeout > 0 {
-                    request = request.timeout(Duration::from_millis(request_timeout));
-                }
-                let response = request.send().map_err(|err| {
-                    if cache_all_errors {
-                        self.create(
-                            &cache_key,
-                            &CachedWithCode::join(500, &format!("{:?}", &err)),
-                            error_cache_duration,
-                        );
-                    }
-                    err
-                })?;
-                let status_code = response.status().as_u16();
-
-                // Early return and no caching if the error is a 500 or greater
-                let is_server_error = response.status().is_server_error();
-                if !cache_all_errors && is_server_error {
-                    return Err(ApiError::from_backend_error(
-                        42,
-                        &format!("Got server error for {}", response.text()?),
-                    ));
-                }
-
-                let is_client_error = response.status().is_client_error();
-                let raw_data = response.text()?;
-
-                if is_client_error || is_server_error {
-                    self.create(
-                        &cache_key,
-                        &CachedWithCode::join(status_code, &raw_data),
-                        error_cache_duration,
-                    );
-                    Err(ApiError::from_backend_error(status_code, &raw_data))
-                } else {
-                    self.create(
-                        &cache_key,
-                        &CachedWithCode::join(status_code, &raw_data),
-                        cache_duration,
-                    );
-                    Ok(raw_data.to_string())
-                }
-            }
-        }
-    }
-
-    fn request_cached(
-        &self,
-        client: &reqwest::blocking::Client,
-        url: &str,
-        cache_duration: usize,
-        error_cache_duration: usize,
-    ) -> ApiResult<String> {
-        self.request_cached_advanced(client, url, cache_duration, error_cache_duration, false, 0)
-    }
-}
-
-impl<T: Cache + ?Sized> CacheExt for T {}
 
 fn pipeline_delete(con: &redis::Connection, keys: Iter<String>) {
     let pipeline = &mut pipe();
